switched from zsh to fish
zsh cabal completion is extremely annoying, preventing path completion.
Trying to report this to zsh is also a pain, because they seem to be
using only mail for bug reports and sourceforge for hosting zsh.

fish always seemed interesting, so this triggered me to switch to it,
and I'm very happy with fish and its plugins.  Goodbye zsh, you were
great!